# api/docs/api/openapi.yaml
openapi: 3.0.0
components:
...
  schemas:
    PlayerRole:
      enum:
        - Initiator
        - Duelist
        - Controller
        - Sentinel
        - Flex
        - IGL
      type: string
...
    MatchApiModel:
      properties:
        date:
          type: string
        tournament_id:
          type: number
          format: double
        team1_id:
          type: number
          format: double
        team2_id:
          type: number
          format: double
        type:
          type: string
        team1_score:
          type: number
          format: double
        team2_score:
          type: number
          format: double
        included_on_standings:
          type: boolean
        started:
          type: boolean
        finished:
          type: boolean
        winner_id:
          type: number
          format: double
        id:
          type: number
          format: double
      required:
        - date
        - tournament_id
        - team1_id
        - team2_id
        - type
        - team1_score
        - team2_score
        - included_on_standings
        - started
        - finished
      type: object
      additionalProperties: false
    ItemsWithPagination_MatchApiModel_:
      properties:
        items:
          items:
            $ref: "#/components/schemas/MatchApiModel"
          type: array
        total:
          type: number
          format: double
      required:
        - items
        - total
      type: object
      additionalProperties: false
...
    PlayerAttributesApiModel:
      properties:
        clutch:
          type: number
          format: double
        awareness:
          type: number
          format: double
        aim:
          type: number
          format: double
        positioning:
          type: number
          format: double
        game_reading:
          type: number
          format: double
        resilience:
          type: number
          format: double
        confidence:
          type: number
          format: double
        strategy:
          type: number
          format: double
        adaptability:
          type: number
          format: double
        communication:
          type: number
          format: double
        unpredictability:
          type: number
          format: double
        game_sense:
          type: number
          format: double
        decision_making:
          type: number
          format: double
        rage_fuel:
          type: number
          format: double
        teamwork:
          type: number
          format: double
        utility_usage:
          type: number
          format: double
      required:
        - clutch
        - awareness
        - aim
        - positioning
        - game_reading
        - resilience
        - confidence
        - strategy
        - adaptability
        - communication
        - unpredictability
        - game_sense
        - decision_making
        - rage_fuel
        - teamwork
        - utility_usage
      type: object
      additionalProperties: false
    PlayerApiModel:
      properties:
        nickname:
          type: string
        full_name:
          type: string
        age:
          type: number
          format: double
        country:
          type: string
        team_id:
          type: number
          format: double
        role:
          $ref: "#/components/schemas/PlayerRole"
        player_attributes:
          $ref: "#/components/schemas/PlayerAttributesApiModel"
        id:
          type: number
          format: double
      required:
        - nickname
        - full_name
        - age
        - country
        - team_id
        - role
        - player_attributes
      type: object
      additionalProperties: false
...
    ItemsWithPagination_PlayerApiModel_:
      properties:
        items:
          items:
            $ref: "#/components/schemas/PlayerApiModel"
          type: array
        total:
          type: number
          format: double
      required:
        - items
        - total
      type: object
      additionalProperties: false
    AllPlayerStats:
      properties:
        player:
          $ref: "#/components/schemas/PlayerApiModel"
        kda:
          type: number
          format: double
        winrate:
          type: number
          format: double
        mapWinrate:
          type: number
          format: double
        totalMatchesPlayed:
          type: number
          format: double
        totalMatchesWon:
          type: number
          format: double
        totalMatchesLost:
          type: number
          format: double
        totalMapsPlayed:
          type: number
          format: double
        totalMapsWon:
          type: number
          format: double
        totalMapsLost:
          type: number
          format: double
        totalKills:
          type: number
          format: double
        totalDeaths:
          type: number
          format: double
        totalAssists:
          type: number
          format: double
      required:
        - player
        - kda
        - winrate
        - mapWinrate
        - totalMatchesPlayed
        - totalMatchesWon
        - totalMatchesLost
        - totalMapsPlayed
        - totalMapsWon
        - totalMapsLost
        - totalKills
        - totalDeaths
        - totalAssists
      type: object
      additionalProperties: false
    ItemsWithPagination_AllPlayerStats_:
      properties:
        items:
          items:
            $ref: "#/components/schemas/AllPlayerStats"
          type: array
        total:
          type: number
          format: double
      required:
        - items
        - total
      type: object
      additionalProperties: false
    GameMap:
      enum:
        - Bind
        - Haven
        - Split
        - Ascent
        - Fracture
        - Icebox
        - Breeze
        - Sunset
        - Abyss
        - Lotus
        - Pearl
      type: string
    GameApiModel:
      properties:
        date:
          type: string
        map:
          $ref: "#/components/schemas/GameMap"
        match_id:
          type: number
          format: double
        included_on_standings:
          type: boolean
        started:
          type: boolean
        finished:
          type: boolean
        id:
          type: number
          format: double
      required:
        - date
        - map
        - match_id
        - included_on_standings
        - started
        - finished
      type: object
      additionalProperties: false
    GameStatsApiModel:
      properties:
        team1_score:
          type: number
          format: double
        team2_score:
          type: number
          format: double
        game_id:
          type: number
          format: double
        team1_id:
          type: number
          format: double
        team2_id:
          type: number
          format: double
        winner_id:
          type: number
          format: double
        id:
          type: number
          format: double
      required:
        - team1_score
        - team2_score
        - game_id
        - team1_id
        - team2_id
      type: object
      additionalProperties: false
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
info:
  title: backend
  version: 0.1.0
  contact: {}
paths:
...
  /players:
    get:
      operationId: getPlayers
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsWithPagination_PlayerApiModel_"
      description: Get all players with optional filtering
      security: []
      parameters:
        - description: Optional team ID to filter by
          in: query
          name: teamId
          required: false
          schema:
            format: double
            type: number
        - description: Maximum number of players to return
          in: query
          name: limit
          required: false
          schema:
            default: 10
            format: double
            type: number
        - description: Number of players to skip
          in: query
          name: offset
          required: false
          schema:
            default: 0
            format: double
            type: number
    post:
      operationId: createPlayer
      responses:
        "201":
          description: Player created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerApiModel"
      description: Creates a new player
      security: []
      parameters: []
      requestBody:
        description: The player data to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerApiModel"
              description: The player data to create
  /players/stats:
    get:
      operationId: getPlayersStats
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsWithPagination_AllPlayerStats_"
      description: Retrieves stats for all players
      security: []
      parameters:
        - description: Maximum number of players to include
          in: query
          name: limit
          required: false
          schema:
            default: 10
            format: double
            type: number
        - description: Number of players to skip
          in: query
          name: offset
          required: false
          schema:
            default: 0
            format: double
            type: number
  /players/{playerId}:
    get:
      operationId: getPlayer
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerApiModel"
      description: Get a specific player by ID
      security: []
      parameters:
        - description: The ID of the player to retrieve
          in: path
          name: playerId
          required: true
          schema:
            format: double
            type: number
    put:
      operationId: updatePlayer
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerApiModel"
      description: Updates an existing player
      security: []
      parameters:
        - description: The ID of the player to update
          in: path
          name: playerId
          required: true
          schema:
            format: double
            type: number
      requestBody:
        description: The player data to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerApiModel"
              description: The player data to update
    delete:
      operationId: deletePlayer
      responses:
        "204":
          description: No content
      description: Deletes a player
      security: []
      parameters:
        - description: The ID of the player to delete
          in: path
          name: playerId
          required: true
          schema:
            format: double
            type: number
  /players/{playerId}/stats:
    get:
      operationId: getPlayerStats
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllPlayerStats"
      description: Retrieves stats for a specific player
      security: []
      parameters:
        - description: The ID of the player to retrieve stats for
          in: path
          name: playerId
          required: true
          schema:
            format: double
            type: number
  /players/bulk:
    post:
      operationId: createPlayersBulk
      responses:
        "201":
          description: Players created successfully
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/PlayerApiModel"
                type: array
      description: Creates multiple players from a bulk upload
      security: []
      parameters: []
      requestBody:
        description: Array of player data to create
        required: true
        content:
          application/json:
            schema:
              items:
                $ref: "#/components/schemas/PlayerApiModel"
              type: array
              description: Array of player data to create
  /matches/{matchId}:
    get:
      operationId: getMatch
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchApiModel"
      description: Retrieves a specific match by its ID
      security: []
      parameters:
        - description: The ID of the match to retrieve
          in: path
          name: matchId
          required: true
          schema:
            format: double
            type: number
  /matches:
    get:
      operationId: getMatches
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsWithPagination_MatchApiModel_"
      description: Retrieves all matches with optional filtering
      security: []
      parameters:
        - description: Filter matches by tournament ID
          in: query
          name: tournamentId
          required: false
          schema:
            format: double
            type: number
        - description: Maximum number of matches to return
          in: query
          name: limit
          required: false
          schema:
            default: 10
            format: double
            type: number
        - description: Number of matches to skip
          in: query
          name: offset
          required: false
          schema:
            default: 0
            format: double
            type: number
  /games/{gameId}:
    get:
      operationId: getGame
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameApiModel"
      description: Retrieves a specific game by its ID
      security: []
      parameters:
        - description: The ID of the game to retrieve
          in: path
          name: gameId
          required: true
          schema:
            format: double
            type: number
  /games/{gameId}/play:
    post:
      operationId: playGame
      responses:
        "200":
          description: Game played successfully
      description: Plays a game and completes it with simulated results
      security: []
      parameters:
        - description: The ID of the game to play
          in: path
          name: gameId
          required: true
          schema:
            format: double
            type: number
  /games/match/{matchId}:
    get:
      operationId: getGamesByMatch
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/GameApiModel"
                type: array
      description: Retrieves games for a specific match
      security: []
      parameters:
        - description: The ID of the match to retrieve games for
          in: path
          name: matchId
          required: true
          schema:
            format: double
            type: number
  /games/{gameId}/stats:
    get:
      operationId: getGameStats
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameStatsApiModel"
      description: Retrieves the stats for a specific game
      security: []
      parameters:
        - description: The ID of the game to retrieve stats for
          in: path
          name: gameId
          required: true
          schema:
            format: double
            type: number
servers:
  - url: /api